fix(prettier): respect prettier config field in "package.json"
