+++ lua/conform/util.lua
@@ -1,3 +1,5 @@
+local log = require("conform.log")
+
 local M = {}
 
 ---Find a command in node_modules
@@ -244,6 +246,62 @@ M.shell_build_argv = function(cmd)
 
   table.insert(argv, cmd)
   return argv
+end
+
+---@param file string
+---@return nil|table
+local function read_json(file)
+  local f = io.open(file, "r")
+  if not f then
+    log.error("Unable to open file %s", file)
+    return nil
+  end
+
+  local file_content = f:read("*all") -- Read entire file contents
+  f:close()
+
+  local ok, json = pcall(vim.json.decode, file_content)
+  if not ok then
+    log.error("Unable to parse json file %s", file)
+    return nil
+  end
+
+  return json
+end
+
+-- @param self conform.FormatterConfig
+-- @param ctx conform.Context|conform.RangeContext
+-- @return string[]|nil args the arguments for setting a `prettier` parser if they exist in the options, nil otherwise
+M.prettier_cwd = function(self, ctx)
+  local config_file_names = {
+    -- https://prettier.io/docs/en/configuration.html
+    ".prettierrc",
+    ".prettierrc.json",
+    ".prettierrc.yml",
+    ".prettierrc.yaml",
+    ".prettierrc.json5",
+    ".prettierrc.js",
+    ".prettierrc.cjs",
+    ".prettierrc.mjs",
+    ".prettierrc.toml",
+    "prettier.config.js",
+    "prettier.config.cjs",
+    "prettier.config.mjs",
+  }
+
+  return vim.fs.root(ctx.dirname, function(name, path)
+    if vim.tbl_contains(config_file_names, name) then
+      return true
+    end
+
+    if name == "package.json" then
+      local full_path = vim.fs.joinpath(path, name)
+      local package_data = read_json(full_path)
+      return package_data and package_data.prettier and true or false
+    end
+
+    return false
+  end)
 end
 
 return M
