+++ tests/util/prettier_spec.lua
@@ -1,11 +1,11 @@
 require("plenary.async").tests.add_to_env()
 
-local conform = require("conform")
 local test_util = require("tests.test_util")
+local util = require("conform.util")
 
-local TMP_DIR = "./tmp/formatters/prettierd/"
+local TMP_DIR = "./tmp/formatters/prettier/"
 
-describe("formatters/prettierd", function()
+describe("util/prettier", function()
   before_each(function()
     vim.fn.mkdir(TMP_DIR, "p")
   end)
@@ -22,11 +22,9 @@ describe("formatters/prettierd", function()
       local jsfile = vim.fs.joinpath(TMP_DIR, "some.js")
       vim.fn.writefile({ "" }, jsfile)
 
-      vim.cmd("e " .. jsfile)
+      local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-      local info = conform.get_formatter_info("prettierd")
-
-      assert.equal(nil, info.cwd)
+      assert.equal(nil, cwd)
     end)
 
     describe("config file", function()
@@ -35,11 +33,9 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(TMP_DIR, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        local info = conform.get_formatter_info("prettierd")
-
-        assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), info.cwd)
+        assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), cwd)
       end)
 
       it("looks up recursively", function()
@@ -52,11 +48,9 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(nested_dir, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        local info = conform.get_formatter_info("prettierd")
-
-        assert.equal(vim.fn.fnamemodify(TMP_DIR, ":p:h"), info.cwd)
+        assert.equal(vim.fn.fnamemodify(TMP_DIR, ":p:h"), cwd)
       end)
     end)
 
@@ -66,16 +60,14 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(TMP_DIR, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
-
         local log = {}
         require("conform.log").set_handler(function(text)
           table.insert(log, text)
         end)
 
-        local info = conform.get_formatter_info("prettierd")
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        assert.equal(nil, info.cwd)
+        assert.equal(nil, cwd)
 
         assert.is_true(#log == 1)
 
@@ -87,11 +79,9 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(TMP_DIR, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        local info = conform.get_formatter_info("prettierd")
-
-        assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), info.cwd)
+        assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), cwd)
       end)
 
       -- test it explicitly just for a future traveler's clarity
@@ -103,11 +93,9 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(TMP_DIR, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        local info = conform.get_formatter_info("prettierd")
-
-        assert.equal(nil, info.cwd)
+        assert.equal(nil, cwd)
       end)
 
       it("looks up recursively", function()
@@ -120,11 +108,9 @@ describe("formatters/prettierd", function()
         local jsfile = vim.fs.joinpath(nested_dir, "some.js")
         vim.fn.writefile({ "" }, jsfile)
 
-        vim.cmd("e " .. jsfile)
+        local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-        local info = conform.get_formatter_info("prettierd")
-
-        assert.equal(vim.fn.fnamemodify(TMP_DIR, ":p:h"), info.cwd)
+        assert.equal(vim.fn.fnamemodify(TMP_DIR, ":p:h"), cwd)
       end)
     end)
 
@@ -138,11 +124,9 @@ describe("formatters/prettierd", function()
       local jsfile = vim.fs.joinpath(nested_dir, "some.js")
       vim.fn.writefile({ "" }, jsfile)
 
-      vim.cmd("e " .. jsfile)
+      local cwd = util.prettier_cwd({}, { dirname = vim.fn.fnamemodify(jsfile, ":p:h") })
 
-      local info = conform.get_formatter_info("prettierd")
-
-      assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), info.cwd)
+      assert.equal(vim.fn.fnamemodify(jsfile, ":p:h"), cwd)
     end)
   end)
 end)
